requirements.txt
aiogram
python-dotenv
requests
httpx
